--- filestore/src/main/kotlin/dev/openrune/cache/Constants.kt
@@ -1,9 +1,0 @@
-package dev.openrune.cache
-
-import com.displee.cache.CacheLibrary
-import dev.openrune.cache.tools.Builder
-
-object Constants {
-    lateinit var builder : Builder
-    lateinit var library: CacheLibrary
-}
