+++ filestore/src/main/kotlin/dev/openrune/cache/util/FileUtil.kt
@@ -1,18 +1,19 @@
 package dev.openrune.cache.util
 
-import dev.openrune.cache.Constants.builder
 import java.io.File
 
 object FileUtil {
 
+    lateinit var cacheLocation : File
+
     fun getTemp() : File {
-        val file = File(builder.cacheLocation,"/temp/")
+        val file = File(cacheLocation,"/temp/")
         if(!file.exists()) file.mkdirs()
         return file
     }
 
     fun getTempDir(dir : String) : File {
-        val file = File(builder.cacheLocation,"/temp/${dir}/")
+        val file = File(cacheLocation,"/temp/${dir}/")
         if(!file.exists()) file.mkdirs()
         return file
     }
