+++ tools/build.gradle.kts
@@ -11,6 +11,8 @@ version = "1.2.4"
 dependencies {
     implementation(project(":filestore"))
     implementation("io.netty:netty-buffer:5.0.0.Alpha2")
+    implementation("dev.openrune:js5server:1.0.2")
+    implementation("net.lingala.zip4j:zip4j:2.11.5")
 }
 
 configure<JavaPluginExtension> {
