+++ tools/src/main/kotlin/dev/openrune/cache/tools/CacheTool.kt
@@ -1,13 +1,11 @@
 package dev.openrune.cache.tools
 
 import com.displee.cache.CacheLibrary
-import dev.openrune.cache.Constants.builder
-import dev.openrune.cache.Constants.library
-import dev.openrune.cache.DownloadOSRS
-import dev.openrune.cache.tools.Builder
-import dev.openrune.cache.tools.DEFAULT_PATH
-import dev.openrune.cache.tools.TaskType
+import dev.openrune.cache.tools.CacheTool.Constants.builder
+import dev.openrune.cache.tools.CacheTool.Constants.library
+import dev.openrune.cache.tools.tasks.TaskType
 import dev.openrune.cache.util.FileUtil
+import dev.openrune.cache.util.FileUtil.cacheLocation
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.jire.js5server.Js5Server
 import java.io.File
@@ -15,15 +13,19 @@ import kotlin.system.exitProcess
 
 class CacheTool(configs: Builder) {
 
+    object Constants {
+        lateinit var builder : Builder
+        lateinit var library: CacheLibrary
+    }
+
     private val logger = KotlinLogging.logger {}
 
     init {
         builder = configs
+        cacheLocation = configs.cacheLocation
     }
 
     fun initialize() {
-
-        println(builder.toString())
 
         if (builder.cacheLocation == DEFAULT_PATH) logger.info { "Using Default path of [${DEFAULT_PATH.absolutePath}]" }
 
@@ -31,7 +33,8 @@ class CacheTool(configs: Builder) {
             TaskType.RUN_JS5 -> Js5Server.init(
                 builder.cacheLocation.absolutePath,
                 builder.js5Ports.toIntArray(), builder.cacheRevision,
-                builder.supportPrefetch)
+                builder.supportPrefetch
+            )
             TaskType.BUILD -> buildCache()
             TaskType.FRESH_INSTALL -> {
                 DownloadOSRS.init()
@@ -62,7 +65,7 @@ class CacheTool(configs: Builder) {
 
     fun runPacking() {
 
-        library?.let {
+        library.let {
             builder.extraTasks.forEach { task ->
                 task.init(it)
             }
@@ -80,8 +83,6 @@ class CacheTool(configs: Builder) {
                 }
 
             FileUtil.getTemp().deleteRecursively()
-        } ?: run {
-            error("Unable to load the cache")
         }
 
     }
