+++ tools/src/main/kotlin/dev/openrune/cache/tools/tasks/CacheTask.kt
@@ -1,4 +1,4 @@
-package dev.openrune.cache.tools
+package dev.openrune.cache.tools.tasks
 
 import com.displee.cache.CacheLibrary
 import dev.openrune.cache.filestore.Cache
