+++ tools/src/main/kotlin/dev/openrune/cache/tools/tasks/TaskType.kt
@@ -1,4 +1,4 @@
-package dev.openrune.cache.tools
+package dev.openrune.cache.tools.tasks
 
 enum class TaskType {
     BUILD,
