+++ tools/src/main/kotlin/dev/openrune/cache/tools/tasks/impl/PackMaps.kt
@@ -5,7 +5,7 @@ import com.google.gson.GsonBuilder
 import dev.openrune.cache.MAPS
 import dev.openrune.cache.filestore.XteaLoader
 import dev.openrune.cache.filestore.logger
-import dev.openrune.cache.tools.CacheTask
+import dev.openrune.cache.tools.tasks.CacheTask
 import dev.openrune.cache.util.decompressGzipToBytes
 import dev.openrune.cache.util.getFiles
 import dev.openrune.cache.util.progress
